Ran logistic regressions on max and mean avg_vel
This commit includes some more preprocessing functions that are for trial-by-trial analysis, as an addition to the total data preprocessing functions that are being used in statistical analysis as of this commit. I've also included logistic regressions using both statsmodels.api as well as SciKitLearn (the logistic regression model from sklearn does not have tuned hyperparameters and just uses a 80-20 train-test split).